Split kfctl_go_tests
Moving to ci

changing the import path

Fixing python path
