+++ testing/kfctl/kfctl_go_test.py
@@ -2,11 +2,9 @@ import logging
 import os
 
 import pytest
-import yaml
+
 from kubeflow.ci import kfctl_go_test_utils as kfctl_util
 from kubeflow.testing import util
-from retrying import retry
-
 
 def test_build_kfctl_go(app_path, project, use_basic_auth, use_istio, config_path):
   """Test building and deploying Kubeflow.
